llm_envs: collapse PreferenceGym + SFTGym into one descriptor-configured DatasetEnv
The preference and SFT gyms differed only by their required columns and collate function (SFT
also a response_column) — descriptors, not subtypes — and neither does generation. Introduce
one DatasetEnv(IterablePromptBatchGym) parameterised by (required_columns, collate_builder,
response_column); keep PreferenceGym/SFTGym as thin back-compat subclasses and preference.py/
sft.py as re-export shims. The generic missing-columns assert message is stabilised (sorted
columns); the preference test's match is updated to it. First slice of the LLM env taxonomy
redesign (docs/design/llm-env-taxonomy.md). test_llm_envs.py green (52 passed).
